Switch color functions return type to uint8_t

+++ include/mandelbrot.h
@@ -10,7 +10,7 @@ extern "C" {
 #include <math.h>
 #include <stdint.h>
 
-typedef int (*color_mapping_func)(int, int);
+typedef uint8_t (*color_mapping_func)(int, int);
 
 // calculeaza f꜀(z) = z² + c
 void mandelbrot_quadratic(double z_real, double z_im, double c_real, double c_im, double *rez_real, double *rez_im);
@@ -54,14 +54,14 @@ uint32_t* deseneaza_mandelbrot(
     const char* nume_poza, int inaltime_poza, int latime_poza,
     double top_left_coord_real, double top_left_coord_imaginar, double pixel_width,
     int num_iters, double rotate_degrees, double brightness,
-    int (*red_mapping_func)(int, int), int (*green_mapping_func)(int, int), int (*blue_mapping_func)(int, int)
+    uint8_t (*red_mapping_func)(int, int), uint8_t (*green_mapping_func)(int, int), uint8_t (*blue_mapping_func)(int, int)
 );
 
 uint32_t* mandelbrot_around_center(
     const char* nume_poza, int inaltime_poza, int latime_poza,
     double center_coord_real, double center_coord_imaginar, double radius,
     int num_iters, double rotate_degrees, double brightness,
-    int (*red_mapping_func)(int, int), int (*green_mapping_func)(int, int), int (*blue_mapping_func)(int, int)
+    uint8_t (*red_mapping_func)(int, int), uint8_t (*green_mapping_func)(int, int), uint8_t (*blue_mapping_func)(int, int)
 );
 
 #ifdef __cplusplus
